(fix/refactor): Fix after refactoring
