+++ mysite/fsm/mixins.py
@@ -146,291 +146,291 @@ class JSONBlobMixin(object):
             raise AttributeError('JSON data has no attr %s' % attr)
 
 
-class ChatMixin(object):
-    """
-    Allow to create message based on current FSM node type.
-    """
-    def node_name_is_one_of(self, *args):
-        return self.name in args
-
-    def get_message(self, chat, current=None, message=None):
-        message = chat.state.fsmNode._plugin.get_message(chat, message, current)
-        return message
-
-        #
-        #   This code is not used anymore, I leave it here just for debug purposes.
-        #
-        # if self.node_name_is_one_of('LESSON'):
-        #     input_type = 'custom'
-        #     kind = next_lesson.lesson.kind
-        #     try:
-        #         if is_additional:
-        #             raise UnitLesson.DoesNotExist
-        #         unitStatus = chat.state.get_data_attr('unitStatus')
-        #         next_ul = unitStatus.unit.unitlesson_set.get(order=unitStatus.order+1)
-        #         if next_ul and next_ul.lesson.kind in SIMILAR_KINDS and kind in SIMILAR_KINDS:
-        #             input_type = 'options'
-        #             kind = 'button'
-        #     except UnitLesson.DoesNotExist:
-        #         pass
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='unitlesson',
-        #                     content_id=next_lesson.id,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     input_type=input_type,
-        #                     kind=kind,
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('ASK'):
-        #     _data = {
-        #         'contenttype': 'unitlesson',
-        #         'content_id': next_lesson.id,
-        #         'chat': chat,
-        #         'owner': chat.user,
-        #         'input_type': 'custom',
-        #         'kind': next_lesson.lesson.kind,
-        #         'is_additional': is_additional
-        #     }
-        #     if not self.fsm.fsm_name_is_one_of('live_chat'):
-        #         message = Message.objects.get_or_create(**_data)[0]
-        #     else:
-        #         message = Message(**_data)
-        #         message.save()
-        # if self.node_name_is_one_of('GET_ANSWER'):
-        #     answer = current.get_answers().first()
-        #     _data = {
-        #         'contenttype': 'response',
-        #         'input_type': 'text',
-        #         'lesson_to_answer': current,
-        #         'chat': chat,
-        #         'owner': chat.user,
-        #         'kind': 'response',
-        #         'userMessage': True,
-        #         'is_additional': is_additional
-        #     }
-        #     if not self.fsm.fsm_name_is_one_of('live_chat'):
-        #         message = Message.objects.get_or_create(**_data)[0]
-        #     else:
-        #         message = Message(**_data)
-        #         message.save()
-        # if self.node_name_is_one_of('CONFIDENCE'):
-        #     # current here is Response instance
-        #     if isinstance(current, Response):
-        #         response_to_chk = current
-        #         answer = current.unitLesson.get_answers().first()
-        #     else:
-        #         response_to_chk = message.response_to_check
-        #         if not message.lesson_to_answer:
-        #             answer = message.response_to_check.unitLesson.get_answers().first()
-        #         else:
-        #             answer = message.lesson_to_answer.get_answers().first()
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='unitlesson',
-        #                     response_to_check=response_to_chk,
-        #                     input_type='custom',
-        #                     text=self.title,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     kind=answer.kind,
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('GET_CONFIDENCE'):
-        #     _data = dict(
-        #         contenttype='response',
-        #         content_id=message.response_to_check.id,
-        #         input_type='options',
-        #         chat=chat,
-        #         owner=chat.user,
-        #         kind='response',
-        #         userMessage=True,
-        #         is_additional=is_additional,
-        #     )
-        #     if not self.fsm.fsm_name_is_one_of('live_chat'):
-        #         message = Message.objects.get_or_create(**_data)[0]
-        #     else:
-        #         message = Message(**_data)
-        #         message.save()
-        # if self.node_name_is_one_of("WAIT_ASSESS"):
-        #     if isinstance(current, Response):
-        #         resp_to_chk = current
-        #     else:
-        #         resp_to_chk = message.response_to_check
-        #     message = Message.objects.get_or_create(
-        #         chat=chat,
-        #         text=self.title,
-        #         kind='button',
-        #         response_to_check=resp_to_chk,
-        #         is_additional=is_additional,
-        #         owner=chat.user,
-        #     )[0]
-        #
-        # if self.node_name_is_one_of('ASSESS'):
-        #     # current here is Response instance
-        #     if isinstance(current, Response):
-        #         response_to_chk = current
-        #         answer = current.unitLesson.get_answers().first()
-        #     else:
-        #         response_to_chk = message.response_to_check
-        #         if not message.lesson_to_answer:
-        #             answer = message.response_to_check.unitLesson.get_answers().first()
-        #         else:
-        #             answer = message.lesson_to_answer.get_answers().first()
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='unitlesson',
-        #                     response_to_check=response_to_chk,
-        #                     input_type='custom',
-        #                     content_id=answer.id,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     kind=answer.kind,
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('GET_ASSESS'):
-        #     _data = dict(
-        #         contenttype='response',
-        #         content_id=message.response_to_check.id,
-        #         input_type='options',
-        #         chat=chat,
-        #         owner=chat.user,
-        #         kind='response',
-        #         userMessage=True,
-        #         is_additional=is_additional
-        #     )
-        #     # if not self.fsm.name == 'live_chat':
-        #     #     message = Message.objects.get_or_create(**_data)[0]
-        #     # else:
-        #     message = Message(**_data)
-        #     message.save()
-        # if self.node_name_is_one_of('STUDENTERROR'):
-        #     resolve_message = Message.objects.get(
-        #                     contenttype='unitlesson',
-        #                     content_id=next_lesson.id,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     input_type='custom',
-        #                     kind='message',
-        #                     timestamp__isnull=True,
-        #                     is_additional=True)
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='unitlesson',
-        #                     content_id=resolve_message.student_error.errorModel.id,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     student_error=resolve_message.student_error,
-        #                     input_type='options',
-        #                     kind='button',
-        #                     is_additional=True)[0]
-        # if self.node_name_is_one_of('RESOLVE'):
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='unitlesson',
-        #                     content_id=next_lesson.id,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     input_type='custom',
-        #                     kind='message',
-        #                     timestamp__isnull=True,
-        #                     is_additional=True)[0]
-        # if self.node_name_is_one_of('MESSAGE_NODE'):
-        #     message = Message.objects.get_or_create(
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     text=chat.state.fsmNode.title,
-        #                     student_error=message.student_error,
-        #                     input_type='custom',
-        #                     kind='message',
-        #                     is_additional=True)[0]
-        # if self.node_name_is_one_of('END', 'IF_RESOURCES'):
-        #     if not self.help:
-        #         text = self.get_help(chat.state, request=None)
-        #     else:
-        #         text = self.help
-        #     message = Message.objects.create(
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     text=text,
-        #                     input_type='custom',
-        #                     kind='message',
-        #                     is_additional=True)
-        # if self.node_name_is_one_of('GET_RESOLVE'):
-        #         message = Message.objects.create(
-        #                     contenttype='unitlesson',
-        #                     content_id=next_lesson.id,
-        #                     input_type='options',
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     student_error=message.student_error,
-        #                     kind='response',
-        #                     userMessage=True,
-        #                     is_additional=is_additional)
-        # if self.node_name_is_one_of('ERRORS'):
-        #     message = Message.objects.get_or_create(
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     text='''Below are some common misconceptions. '''
-        #                          '''Select one or more that is similar to your reasoning.''',
-        #                     kind='message',
-        #                     input_type='custom',
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('GET_ERRORS'):
-        #     uniterror = UnitError.get_by_message(message)
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='uniterror',
-        #                     content_id=uniterror.id,
-        #                     input_type='options',
-        #                     chat=chat,
-        #                     kind='uniterror',
-        #                     owner=chat.user,
-        #                     userMessage=False,
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('TITLE'):
-        #     divider = ChatDivider(text=next_lesson.lesson.title,
-        #                           unitlesson=next_lesson)
-        #     divider.save()
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='chatdivider',
-        #                     content_id=divider.id,
-        #                     input_type='custom',
-        #                     type='breakpoint',
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     kind='message',
-        #                     is_additional=is_additional)[0]
-        # if self.node_name_is_one_of('START_MESSAGE'):
-        #     message = Message.objects.create(
-        #                     input_type='options',
-        #                     text=self.title,
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     kind='button',
-        #                     is_additional=is_additional)
-        # if self.node_name_is_one_of('DIVIDER'):
-        #     divider = ChatDivider(text=self.title)
-        #     divider.save()
-        #     message = Message.objects.get_or_create(
-        #                     contenttype='chatdivider',
-        #                     content_id=divider.id,
-        #                     input_type='custom',
-        #                     type='breakpoint',
-        #                     chat=chat,
-        #                     owner=chat.user,
-        #                     kind='message',
-        #                     is_additional=is_additional)[0]
-        #
-        # if self.node_name_is_one_of('START') and self.fsm.fsm_name_is_one_of('live_chat'):
-        #     message = Message.objects.get_or_create(
-        #         chat=chat,
-        #         text=self.title,
-        #         kind='button',
-        #         is_additional=True,
-        #         owner=chat.user,
-        #     )[0]
-        #
-        # # wait for RECYCLE node and  any node starting from WAIT_ except WAIT_ASSESS
-        # if is_wait_node(self.name):
-        #     lookup = dict(
-        #         chat=chat,
-        #         text=self.title,
-        #         kind='button',
-        #         is_additional=False,
-        #         owner=chat.user
-        #     )
-        #     message = Message.objects.get_or_create(**lookup)[0]
-        # return message
+# class ChatMixin(object):
+#     """
+#     Allow to create message based on current FSM node type.
+#     """
+#     def node_name_is_one_of(self, *args):
+#         return self.name in args
+#
+#     def get_message(self, chat, current=None, message=None):
+#         message = chat.state.fsmNode._plugin.get_message(chat, message, current)
+#         return message
+#
+#
+#           This code is not used anymore, I leave it here just for debug purposes.
+#
+#         if self.node_name_is_one_of('LESSON'):
+#             input_type = 'custom'
+#             kind = next_lesson.lesson.kind
+#             try:
+#                 if is_additional:
+#                     raise UnitLesson.DoesNotExist
+#                 unitStatus = chat.state.get_data_attr('unitStatus')
+#                 next_ul = unitStatus.unit.unitlesson_set.get(order=unitStatus.order+1)
+#                 if next_ul and next_ul.lesson.kind in SIMILAR_KINDS and kind in SIMILAR_KINDS:
+#                     input_type = 'options'
+#                     kind = 'button'
+#             except UnitLesson.DoesNotExist:
+#                 pass
+#             message = Message.objects.get_or_create(
+#                             contenttype='unitlesson',
+#                             content_id=next_lesson.id,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             input_type=input_type,
+#                             kind=kind,
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('ASK'):
+#             _data = {
+#                 'contenttype': 'unitlesson',
+#                 'content_id': next_lesson.id,
+#                 'chat': chat,
+#                 'owner': chat.user,
+#                 'input_type': 'custom',
+#                 'kind': next_lesson.lesson.kind,
+#                 'is_additional': is_additional
+#             }
+#             if not self.fsm.fsm_name_is_one_of('live_chat'):
+#                 message = Message.objects.get_or_create(**_data)[0]
+#             else:
+#                 message = Message(**_data)
+#                 message.save()
+#         if self.node_name_is_one_of('GET_ANSWER'):
+#             answer = current.get_answers().first()
+#             _data = {
+#                 'contenttype': 'response',
+#                 'input_type': 'text',
+#                 'lesson_to_answer': current,
+#                 'chat': chat,
+#                 'owner': chat.user,
+#                 'kind': 'response',
+#                 'userMessage': True,
+#                 'is_additional': is_additional
+#             }
+#             if not self.fsm.fsm_name_is_one_of('live_chat'):
+#                 message = Message.objects.get_or_create(**_data)[0]
+#             else:
+#                 message = Message(**_data)
+#                 message.save()
+#         if self.node_name_is_one_of('CONFIDENCE'):
+#             # current here is Response instance
+#             if isinstance(current, Response):
+#                 response_to_chk = current
+#                 answer = current.unitLesson.get_answers().first()
+#             else:
+#                 response_to_chk = message.response_to_check
+#                 if not message.lesson_to_answer:
+#                     answer = message.response_to_check.unitLesson.get_answers().first()
+#                 else:
+#                     answer = message.lesson_to_answer.get_answers().first()
+#             message = Message.objects.get_or_create(
+#                             contenttype='unitlesson',
+#                             response_to_check=response_to_chk,
+#                             input_type='custom',
+#                             text=self.title,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             kind=answer.kind,
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('GET_CONFIDENCE'):
+#             _data = dict(
+#                 contenttype='response',
+#                 content_id=message.response_to_check.id,
+#                 input_type='options',
+#                 chat=chat,
+#                 owner=chat.user,
+#                 kind='response',
+#                 userMessage=True,
+#                 is_additional=is_additional,
+#             )
+#             if not self.fsm.fsm_name_is_one_of('live_chat'):
+#                 message = Message.objects.get_or_create(**_data)[0]
+#             else:
+#                 message = Message(**_data)
+#                 message.save()
+#         if self.node_name_is_one_of("WAIT_ASSESS"):
+#             if isinstance(current, Response):
+#                 resp_to_chk = current
+#             else:
+#                 resp_to_chk = message.response_to_check
+#             message = Message.objects.get_or_create(
+#                 chat=chat,
+#                 text=self.title,
+#                 kind='button',
+#                 response_to_check=resp_to_chk,
+#                 is_additional=is_additional,
+#                 owner=chat.user,
+#             )[0]
+#
+#         if self.node_name_is_one_of('ASSESS'):
+#             # current here is Response instance
+#             if isinstance(current, Response):
+#                 response_to_chk = current
+#                 answer = current.unitLesson.get_answers().first()
+#             else:
+#                 response_to_chk = message.response_to_check
+#                 if not message.lesson_to_answer:
+#                     answer = message.response_to_check.unitLesson.get_answers().first()
+#                 else:
+#                     answer = message.lesson_to_answer.get_answers().first()
+#             message = Message.objects.get_or_create(
+#                             contenttype='unitlesson',
+#                             response_to_check=response_to_chk,
+#                             input_type='custom',
+#                             content_id=answer.id,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             kind=answer.kind,
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('GET_ASSESS'):
+#             _data = dict(
+#                 contenttype='response',
+#                 content_id=message.response_to_check.id,
+#                 input_type='options',
+#                 chat=chat,
+#                 owner=chat.user,
+#                 kind='response',
+#                 userMessage=True,
+#                 is_additional=is_additional
+#             )
+#             # if not self.fsm.name == 'live_chat':
+#             #     message = Message.objects.get_or_create(**_data)[0]
+#             # else:
+#             message = Message(**_data)
+#             message.save()
+#         if self.node_name_is_one_of('STUDENTERROR'):
+#             resolve_message = Message.objects.get(
+#                             contenttype='unitlesson',
+#                             content_id=next_lesson.id,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             input_type='custom',
+#                             kind='message',
+#                             timestamp__isnull=True,
+#                             is_additional=True)
+#             message = Message.objects.get_or_create(
+#                             contenttype='unitlesson',
+#                             content_id=resolve_message.student_error.errorModel.id,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             student_error=resolve_message.student_error,
+#                             input_type='options',
+#                             kind='button',
+#                             is_additional=True)[0]
+#         if self.node_name_is_one_of('RESOLVE'):
+#             message = Message.objects.get_or_create(
+#                             contenttype='unitlesson',
+#                             content_id=next_lesson.id,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             input_type='custom',
+#                             kind='message',
+#                             timestamp__isnull=True,
+#                             is_additional=True)[0]
+#         if self.node_name_is_one_of('MESSAGE_NODE'):
+#             message = Message.objects.get_or_create(
+#                             chat=chat,
+#                             owner=chat.user,
+#                             text=chat.state.fsmNode.title,
+#                             student_error=message.student_error,
+#                             input_type='custom',
+#                             kind='message',
+#                             is_additional=True)[0]
+#         if self.node_name_is_one_of('END', 'IF_RESOURCES'):
+#             if not self.help:
+#                 text = self.get_help(chat.state, request=None)
+#             else:
+#                 text = self.help
+#             message = Message.objects.create(
+#                             chat=chat,
+#                             owner=chat.user,
+#                             text=text,
+#                             input_type='custom',
+#                             kind='message',
+#                             is_additional=True)
+#         if self.node_name_is_one_of('GET_RESOLVE'):
+#                 message = Message.objects.create(
+#                             contenttype='unitlesson',
+#                             content_id=next_lesson.id,
+#                             input_type='options',
+#                             chat=chat,
+#                             owner=chat.user,
+#                             student_error=message.student_error,
+#                             kind='response',
+#                             userMessage=True,
+#                             is_additional=is_additional)
+#         if self.node_name_is_one_of('ERRORS'):
+#             message = Message.objects.get_or_create(
+#                             chat=chat,
+#                             owner=chat.user,
+#                             text='''Below are some common misconceptions. '''
+#                                  '''Select one or more that is similar to your reasoning.''',
+#                             kind='message',
+#                             input_type='custom',
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('GET_ERRORS'):
+#             uniterror = UnitError.get_by_message(message)
+#             message = Message.objects.get_or_create(
+#                             contenttype='uniterror',
+#                             content_id=uniterror.id,
+#                             input_type='options',
+#                             chat=chat,
+#                             kind='uniterror',
+#                             owner=chat.user,
+#                             userMessage=False,
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('TITLE'):
+#             divider = ChatDivider(text=next_lesson.lesson.title,
+#                                   unitlesson=next_lesson)
+#             divider.save()
+#             message = Message.objects.get_or_create(
+#                             contenttype='chatdivider',
+#                             content_id=divider.id,
+#                             input_type='custom',
+#                             type='breakpoint',
+#                             chat=chat,
+#                             owner=chat.user,
+#                             kind='message',
+#                             is_additional=is_additional)[0]
+#         if self.node_name_is_one_of('START_MESSAGE'):
+#             message = Message.objects.create(
+#                             input_type='options',
+#                             text=self.title,
+#                             chat=chat,
+#                             owner=chat.user,
+#                             kind='button',
+#                             is_additional=is_additional)
+#         if self.node_name_is_one_of('DIVIDER'):
+#             divider = ChatDivider(text=self.title)
+#             divider.save()
+#             message = Message.objects.get_or_create(
+#                             contenttype='chatdivider',
+#                             content_id=divider.id,
+#                             input_type='custom',
+#                             type='breakpoint',
+#                             chat=chat,
+#                             owner=chat.user,
+#                             kind='message',
+#                             is_additional=is_additional)[0]
+#
+#         if self.node_name_is_one_of('START') and self.fsm.fsm_name_is_one_of('live_chat'):
+#             message = Message.objects.get_or_create(
+#                 chat=chat,
+#                 text=self.title,
+#                 kind='button',
+#                 is_additional=True,
+#                 owner=chat.user,
+#             )[0]
+#
+#         # wait for RECYCLE node and  any node starting from WAIT_ except WAIT_ASSESS
+#         if is_wait_node(self.name):
+#             lookup = dict(
+#                 chat=chat,
+#                 text=self.title,
+#                 kind='button',
+#                 is_additional=False,
+#                 owner=chat.user
+#             )
+#             message = Message.objects.get_or_create(**lookup)[0]
+#         return message
